IWorld, ILevel, IDimension
