+++ Obsidian.API/_Interfaces/IDimension.cs
@@ -1,0 +1,6 @@
+namespace Obsidian.API;
+
+public interface IDimension : ILevel
+{
+    public IWorld ParentWorld { get; }
+}
